requirements.txt
streamlit
cryptography.fernet